requirements_model.txt
tensorflow
pandas
keras
numpy
opencv-python
matplotlib
pillow
pillow-simd
scipy
scikit-learn
streamlit
albumentations
tqdm
torch
torchvision
torchaudio